// LIM014-burger-queen/src/app/model/item.ts
export class Item{
  'id': number = 0;
  'title': string;
  'type':string;
  'price': number= 0;
  'quantity': number= 0;
  'completed': boolean= false;
}
/*
export interface IProductsModel {
  products?: ProductDetailModel
}
export interface ProductDetailModel {
  _id?: string;
  name?: string;
  price?: number;
  image?: URL;
  type: string;
  dateEntry: Date;

} */